+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - ✨(back) add model fallback mechanism
 - ✨(back) add celery for running background tasks
 - 🧱(helm) add celery worker and beat deployments
+- ✨(back) add Staan web search tool
 
 ### Changed